+++ packages/react-router-dom-v5-compat/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `react-router-dom-v5-compat`
 
+## 6.4.4
+
+### Patch Changes
+
+- Updated dependencies:
+  - `react-router-dom@6.4.4`
+  - `react-router@6.4.4`
+
 ## 6.4.3
 
 ### Patch Changes